splits api-keys functionality
